feat: TS 변환 기본 세팅
