+++ package.json
@@ -4,14 +4,18 @@
   "description": "",
   "main": "src/app.ts",
   "scripts": {
-    "dev": "ts-node-dev --respawn src/app.ts",
+    "dev": "ts-node-dev --loader ts-node/esm --respawn src/app.ts",
     "build": "tsc",
     "start": "node --enable-source-maps dist/app.js",
+    "generate": "prisma generate",
     "migrate": "prisma migrate dev",
     "studio": "prisma studio"
   },
+  "prisma": {
+    "schema": "src/db/prisma/schema.prisma"
+  },
   "dependencies": {
-    "@prisma/client": "^6.6.0",
+    "@prisma/client": "^6.9.0",
     "bcrypt": "^5.1.1",
     "cookie-parser": "^1.4.7",
     "cors": "^2.8.5",
@@ -22,13 +26,15 @@
     "morgan": "^1.10.0",
     "multer": "^1.4.5-lts.2"
   },
-  "author": "",
-  "license": "ISC",
   "devDependencies": {
     "@types/bcrypt": "^5.0.2",
     "@types/cookie-parser": "^1.4.9",
+    "@types/cors": "^2.8.19",
     "@types/express": "^5.0.3",
-    "@types/jsonwebtoken": "^9.0.9",
+    "@types/jsonwebtoken": "^9.0.10",
+    "@types/morgan": "^1.9.10",
+    "@types/multer": "^1.4.13",
+    "@types/node": "^24.0.3",
     "nodemon": "^3.1.10",
     "ts-node": "^10.9.2",
     "ts-node-dev": "^2.0.0",
